+++ src/http/controllers/portifolios/create-portifolio.e2e-spec.ts
@@ -1,7 +1,7 @@
+import { getPrisma } from '@/http/prisma/prisma'
 import { Server, createServer } from 'http'
 import request from 'supertest'
 import { makePrismaUser } from 'test/factories/make-user'
-import { getPrisma } from '../prisma/prisma'
 
 let server: Server
