+++ src/http/controllers/users/authenticate.e2e-spec.ts
@@ -1,7 +1,7 @@
+import { getPrisma } from '@/http/prisma/prisma'
 import { hash } from 'bcrypt'
 import { Server, createServer } from 'http'
 import request from 'supertest'
-import { getPrisma } from '../prisma/prisma'
 
 let server: Server
